优化结构，抽离详细信息组件
