+++ page/SearchDetail.js
@@ -9,22 +9,13 @@ import {
   TouchableOpacity,
   TouchableNativeFeedback,
   LayoutAnimation,
-  NativeModules,
   PanResponder,
-  Platform,
 } from 'react-native';
-import {Novel, Chapter, Content, Store} from '../domain';
+import {Novel, Chapter} from '../domain';
+import NovelDetail from './commpent/NovelDetail';
 import http from '../request';
 import EvilIcons from 'react-native-vector-icons/EvilIcons';
 import Loading from './commpent/Loading';
-
-const {UIManager} = NativeModules;
-
-if (Platform.OS === 'android') {
-  if (UIManager.setLayoutAnimationEnabledExperimental) {
-    UIManager.setLayoutAnimationEnabledExperimental(true);
-  }
-}
 
 type State = {
   chapters: Array<Chapter>,
@@ -58,8 +49,6 @@ export default class Detail extends React.Component<Props, State> {
       title: props.route.params.novel.title,
     });
     this.renderChapter = this.renderChapter.bind(this);
-    this.renderInfo = this.renderInfo.bind(this);
-    this.hiddenIntro = this.hiddenIntro.bind(this);
     this.read = this.read.bind(this);
 
     let param = {
@@ -90,74 +79,11 @@ export default class Detail extends React.Component<Props, State> {
           loading: false,
         });
       });
-
-    this._panResponder = PanResponder.create({
-      // Ask to be the responder:
-      onStartShouldSetPanResponder: (evt, gestureState) => true,
-      onStartShouldSetPanResponderCapture: (evt, gestureState) => {
-        let {dx, dy} = gestureState;
-        return Math.abs(dx) > 5 || Math.abs(dy) > 5;
-      },
-      onMoveShouldSetPanResponder: (evt, gestureState) => true,
-      onMoveShouldSetPanResponderCapture: (evt, gestureState) => true,
-
-      onPanResponderGrant: (evt, gestureState) => {
-        // The gesture has started. Show visual feedback so the user knows
-        // what is happening!
-        // gestureState.d{x,y} will be set to zero now UIManager
-        if (!this.introHeight) {
-          this.introDoc.measure((x, y, widths, heights, pageX, pageY) => {
-            this.introHeight = heights;
-          });
-        }
-      },
-      onPanResponderMove: (evt, gestureState) => {
-        // The most recent move distance is gestureState.move{X,Y}
-        // The accumulated gesture distance since becoming responder is
-        // gestureState.d{x,y}
-        let dy = gestureState.dy;
-        if (dy < 0 && this.introHeight) {
-          this.nowHeight =
-            this.introHeight + dy < 1 ? 1 : this.introHeight + dy;
-          this.introDoc.setNativeProps({
-            height: this.nowHeight,
-          });
-        }
-      },
-      onPanResponderTerminationRequest: (evt, gestureState) => true,
-      onPanResponderRelease: (evt, gestureState) => {
-        // The user has released all touches while this view is the
-        // responder. This typically means a gesture has succeeded
-        this.introDoc.setNativeProps({
-          height: 'auto',
-        });
-        if (gestureState.dy < -50 && !this.state.hidden) {
-          this.hiddenIntro();
-        }
-      },
-      onPanResponderTerminate: (evt, gestureState) => {
-        // Another component has become the responder, so this gesture
-        // should be cancelled
-      },
-      onShouldBlockNativeResponder: (evt, gestureState) => {
-        // Returns whether this component should block native components from becoming the JS
-        // responder. Returns true by default. Is currently only supported on android.
-        return true;
-      },
-    });
   }
   read(chapter: Chapter) {
     this.props.navigation.navigate('SearchRead', {
       chapter: chapter,
       novel: this.state.novel,
-    });
-  }
-  hiddenIntro() {
-    LayoutAnimation.configureNext(
-      LayoutAnimation.create(100, 'linear', 'opacity'),
-    );
-    this.setState({
-      hidden: !this.state.hidden,
     });
   }
   renderChapter({item}) {
@@ -170,45 +96,27 @@ export default class Detail extends React.Component<Props, State> {
       </TouchableOpacity>
     );
   }
-  renderInfo() {
-    return (
-      <View
-        {...this._panResponder.panHandlers}
-        collapsable={false}
-        style={this.state.hidden ? {height: 0} : {opacity: 1}}
-        ref={doc => (this.introDoc = doc)}>
-        <View style={styles.info}>
-          <Image
-            source={
-              this.state.coverUrl
-                ? {uri: this.state.coverUrl}
-                : require('../images/nopic.jpg')
-            }
-            style={styles.thumbnail}
-          />
-          <View style={styles.rightContainer}>
-            <Text style={styles.infoText}>{this.state.novel.name}</Text>
-            <Text style={styles.infoText}>
-              作者：{this.state.novel.author.name}
-            </Text>
-          </View>
-        </View>
-        <View style={styles.introduce}>
-          <Text style={styles.title}>简介：</Text>
-          <Text>{this.state.novel.introduce}</Text>
-        </View>
-      </View>
-    );
-  }
   render() {
     return (
       <View style={styles.container}>
-        {this.renderInfo()}
+        <NovelDetail
+          cover={
+            this.state.coverUrl
+              ? {uri: this.state.coverUrl}
+              : require('../images/nopic.jpg')
+          }
+          name={this.state.novel.name}
+          author={this.state.novel.author.name}
+          introduce={this.state.novel.introduce}
+          ref={ref => {
+            this.novelDetail = ref;
+          }}
+        />
         <View style={styles.chapterContainer}>
           <View style={styles.ChapterListBox}>
             <TouchableNativeFeedback
               onPress={() => {
-                this.hiddenIntro();
+                this.novelDetail.hiddenIntro();
               }}>
               <View style={styles.chapterListTitle}>
                 <Text style={styles.title}>
@@ -222,11 +130,11 @@ export default class Detail extends React.Component<Props, State> {
           <FlatList
             data={this.state.chapters}
             renderItem={this.renderChapter}
-            keyExtractor={item => item.url}
+            keyExtractor={(item, index) => item.url + index}
             ListEmptyComponent={<Text style={styles.tip}>暂无数据</Text>}
           />
         </View>
-        {this.state.coverUrl ? null : <Loading />}
+        {this.state.loading ? <Loading /> : null}
       </View>
     );
   }
